Create move selector union test
Adding stepping configuration with LA To track the cashed moves

Cash usage introduced 8.33, affects big datasets
