+++ optaplanner-benchmarks/optaplanner-perf-benchmark/src/main/java/org/jboss/qa/brms/performance/localsearch/cloudbalance/moveselector/union/CloudBalancePillarChangeAndSwapMoveSelectorBenchmark.java
@@ -1,4 +1,4 @@
-package org.jboss.qa.brms.performance.localsearch.cloudbalance.moveselector;
+package org.jboss.qa.brms.performance.localsearch.cloudbalance.moveselector.union;
 
 import org.jboss.qa.brms.performance.configuration.MoveSelectorConfigurations;
 import org.openjdk.jmh.annotations.Benchmark;
@@ -7,15 +7,10 @@ import org.optaplanner.examples.cloudbalancing.domain.CloudBalance;
 
 import java.util.List;
 
-public class CloudBalancePillarChangeAndSwapMoveSelectorBenchmark extends AbstractCloudBalanceMoveSelectorBenchmark {
+public class CloudBalancePillarChangeAndSwapMoveSelectorBenchmark extends AbstractCloudBalanceMoveSelectorUnionBenchmark {
     @Override
     public List<MoveSelectorConfig> createMoveSelectorConfigList() {
         return MoveSelectorConfigurations.createMultiPilarSelectorList();
-    }
-    @Override
-    public int getAcceptedCountLimit() {
-        //to check that pillar move selector caches on each new step
-        return 100;
     }
 
     @Benchmark
